Cleanup * combine styled components into one file (popover, shadow) * remove unnecessary width * add Shadow and animation
